refactor(strucutre)
Move service files and rename the manageExtension and util folders
