+++ src/extensionManager/Extension.spec.ts
