+++ src/extensionManager/ExtensionsManager.spec.ts
