+++ src/extensionManager/ExtensionsManager.ts
