+++ src/extensionManager/Feature.ts
